Server/pyproject.toml
[tool.poetry]
name = "gohan-map-backend"
version = "0.1.0"
description = ""
authors = ["karintou8710"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
poethepoet = "^0.24.1"
sqlalchemy = "^2.0.22"
alembic = "^1.12.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
firebase-admin = "^6.2.0"
python-multipart = "^0.0.6"
sqlalchemy-utils = "^0.41.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
mypy = "^1.6.0"
isort = "^5.12.0"
pytest = "^7.4.3"
httpx = "^0.25.1"


[tool.poe.tasks]
serve = "uvicorn main:app --reload"
format = "bash -c 'isort . && black .'"
format-check = "bash -c 'isort --check . && black --check .'"
type = "mypy app/ tests/"

[tool.isort]
profile = "black"

[tool.mypy]
strict = true
ignore_missing_imports = true
plugins = "sqlalchemy.ext.mypy.plugin"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
